move some utils to linearElementEditor.ts
